LVM migration: Check if name is equal to dest_vg
The existence check of the destination volume group is wrong. It
currently checks if there's a VG with a name equal to the source's vg
name instead of checking if it's equal to the dest_vg.

Closes-bug: #1258203
(cherry picked from commit 17912f4c4fba270a15cab47bb68d9e30c34a69b0)

Change-Id: Ia5d4acb24b94c6aa832107c7eb4b6996985af97f